+++ metadata.rb
@@ -10,7 +10,7 @@ long_description 'Installs/Configures Divvy'
 version          '0.2.1'
 
 depends          'mac-app-store', '~> 1.0'
-depends          'macosx_accessibility', '~> 1.0'
+depends          'privacy_services_manager', '~> 1.0'
 depends          'windows', '~> 1.36'
 
 supports         'mac_os_x'