+++ spec/spec_helper.rb
@@ -11,9 +11,9 @@ require 'tmpdir'
 require 'fileutils'
 require_relative '../libraries/matchers'
 require_relative 'support/resource/mac_app_store_app'
-require_relative 'support/resource/macosx_accessibility'
+require_relative 'support/resource/privacy_services_manager'
 require_relative 'support/provider/mac_app_store_app'
-require_relative 'support/provider/macosx_accessibility'
+require_relative 'support/provider/privacy_services_manager'
 
 RSpec.configure do |c|
   c.color = true