+++ test/integration/mac_direct_download/serverspec/localhost/app_spec.rb
@@ -18,11 +18,11 @@ describe 'Divvy app' do
     end
   end
 
-  # TODO: This should pass in Windows too, but Specinfra throws a
-  # NotImplementedError.
-  describe process('Divvy') do
+  # TODO: Using proces('Divvy') requires a fix for Specinfra to not try to use
+  # `ps -C` in OS X.
+  describe command('ps -A -c -o command | grep Divvy') do
     it 'is running' do
-      expect(subject).to be_running
+      expect(subject.stdout.strip).to eq('Divvy')
     end
   end
 end